+++ src/utils/abstract-observer.js
@@ -11,7 +11,7 @@ export default class AbstractObserver {
     this._observers.delete(observer);
   }
 
-  _notify(event, payload, payloadSource) {
-    this._observers.forEach((observer) => observer(event, payload, payloadSource));
+  _notify(event, payload) {
+    this._observers.forEach((observer) => observer(event, payload));
   }
 }
